apps/test-app/app/tests/progress/index.spec.ts
/*---------------------------------------------------------------------------------------------
 * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
import { test, expect } from "#playwright";
import AxeBuilder from "@axe-core/playwright";

test("default", async ({ page }) => {
	await page.goto("/tests/progress");
	const progressbar = page.getByRole("progressbar");
	await expect(progressbar).toBeVisible();
});

test("@visual", async ({ page }) => {
	await page.goto("/tests/progress?visual=true&paused");
	await expect(page.locator("body")).toHaveScreenshot();
});

test.describe("@a11y", () => {
	test("Axe Page Scan", async ({ page }) => {
		await page.goto("/tests/progress");

		const progressbar = page.getByRole("progressbar");
		await expect(progressbar).toBeVisible();

		const axe = new AxeBuilder({ page });
		const accessibilityScan = await axe.analyze();
		expect(accessibilityScan.violations).toEqual([]);
	});
});
